requirements.txt
tensorflow-cpu
numpy
pillow
matplotlib
scikit-image
scipy
flask
gunicorn